# jobportal-docs/apis/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "Register a new user (JobSeeker, Employer, or Admin)",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User registration data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              },
              "examples": {
                "jobseeker": {
                  "summary": "JobSeeker Registration",
                  "value": {
                    "email": "john.doe@example.com",
                    "password": "SecurePass123!",
                    "role": "JobSeeker",
                    "name": "John Doe",
                    "address": "123 Main St, New York, NY 10001",
                    "telephone": "+1-555-123-4567"
                  }
                },
                "employer": {
                  "summary": "Employer Registration",
                  "value": {
                    "email": "hr@techcorp.com",
                    "password": "SecurePass123!",
                    "role": "Employer",
                    "name": "Jane Smith",
                    "address": "456 Business Ave, San Francisco, CA 94105",
                    "telephone": "+1-555-987-6543"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Registration successful"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "summary": "User Registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Login with email and password to get JWT token",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful - JWT token provided",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "user": {
                      "id": "507f1f77bcf86cd799439011",
                      "email": "user@example.com",
                      "role": "JobSeeker"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "User Login",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/applications/{jobId}": {
      "post": {
        "description": "Submit a job application (JobSeeker only)",
        "operationId": "ApplicationsController_apply",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Job ID to apply for",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              },
              "examples": {
                "simpleApplication": {
                  "summary": "Basic Application",
                  "value": {
                    "coverLetter": "I am very interested in this position."
                  }
                },
                "fullApplication": {
                  "summary": "Full Application with Resume",
                  "value": {
                    "resumeUrl": "https://cloudinary.com/...",
                    "coverLetter": "Dear Hiring Manager, I am very interested in the Senior Software Engineer position. With 7 years of experience in React and Node.js, I am confident..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Only JobSeekers can apply"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Apply for Job",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/me": {
      "get": {
        "description": "Retrieve all applications submitted by the authenticated user",
        "operationId": "ApplicationsController_myApplications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Applications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Application"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get My Applications",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/my-jobs": {
      "get": {
        "description": "Get all jobs the authenticated JobSeeker has applied to",
        "operationId": "ApplicationsController_myAppliedJobs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Application"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Only JobSeekers can access"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get My Applied Jobs",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/job/{jobId}": {
      "get": {
        "description": "Get all applications for a specific job",
        "operationId": "ApplicationsController_applicationsForJob",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Job ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Applications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Application"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get Job Applications",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{applicationId}": {
      "delete": {
        "description": "Delete an application (Employer only)",
        "operationId": "ApplicationsController_deleteApplication",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "description": "Application ID to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application deleted successfully"
          },
          "404": {
            "description": "Application not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete Application",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{applicationId}/status": {
      "patch": {
        "description": "Update the status of an application (Employer only)",
        "operationId": "ApplicationsController_updateApplicationStatus",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "description": "Application ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "Applied",
                      "Viewed",
                      "Rejected"
                    ]
                  }
                },
                "required": [
                  "status"
                ]
              },
              "examples": {
                "viewed": {
                  "summary": "Mark as Viewed",
                  "value": {
                    "status": "Viewed"
                  }
                },
                "rejected": {
                  "summary": "Reject Application",
                  "value": {
                    "status": "Rejected"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Application status updated"
          },
          "404": {
            "description": "Application not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update Application Status",
        "tags": [
          "Applications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Job Portal API",
    "description": "A comprehensive API for managing jobs, applications, users, and messaging. Supports multiple user roles: JobSeeker, Employer, and Admin.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "JobSeeker",
              "Employer",
              "Admin"
            ],
            "example": "JobSeeker"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "address": {
            "type": "string",
            "example": "123 Main St, New York, NY"
          }
        },
        "required": [
          "email",
          "password",
          "role",
          "name",
          "address"
        ]
      },
...
      "CreateApplicationDto": {
        "type": "object",
        "properties": {}
      },
      "Application": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "Applied",
              "Viewed",
              "Rejected"
            ],
            "example": "Applied"
          },
          "resumeUrl": {
            "type": "string",
            "example": "https://cloudinary.com/resume.pdf"
          },
          "coverLetter": {
            "type": "string",
            "example": "I am highly motivated..."
          },
          "jobId": {
            "type": "string",
            "example": "65c123abc..."
          },
          "jobSeekerId": {
            "type": "string",
            "example": "65d456def..."
          }
        },
        "required": [
          "status",
          "resumeUrl",
          "coverLetter",
          "jobId",
          "jobSeekerId"
        ]
      },
...
    }
  }
}